Check dryrun

+++ album.go
@@ -251,8 +251,13 @@ func (self *Album) Relocate() error {
 
 		if maykill && RemoveOld {
 			// we can remove the file from the old location
-			if err := os.Remove(info.path); err != nil {
-				self.failed[info.path] = err
+			if DryRun || Verbose {
+				fmt.Printf("# deleting %s\n", info.path)
+			}
+			if !DryRun {
+				if err := os.Remove(info.path); err != nil {
+					self.failed[info.path] = err
+				}
 			}
 		}
 	}
